+++ src/components/TodoList.vue
@@ -1,22 +1,5 @@
 <template>
-  <div class="form-wrapper">
-    <NForm
-      ref="formRef"
-      inline
-      :label-width="80"
-      :model="formValue"
-      :rules="rules"
-      size="large"
-      @submit="handleSubmit"
-    >
-      <n-form-item path="task">
-        <n-input v-model:value="formValue.task" placeholder="Task" />
-      </n-form-item>
-      <n-form-item>
-        <n-button attr-type="submit"> Add </n-button>
-      </n-form-item>
-    </NForm>
-  </div>
+  <AddTaskForm />
   <n-scrollbar x-scrollable>
     <div class="columns-wrapper">
       <div v-for="column in columns" :key="column.value">
@@ -26,10 +9,10 @@
           :list="getTaskByColumn(column.value)"
           group="columns"
           item-key="title"
-          @change="log"
+          @change="onChange"
         >
           <template #item="{ element }">
-            <div class="list-group-item">{{ element.title }}</div>
+            <div>{{ element.title }}</div>
           </template>
         </draggable>
       </div>
@@ -38,26 +21,14 @@
 </template>
 
 <script setup>
-import { useTasksStore } from "../stores/TasksStore";
-import {
-  NForm,
-  NFormItem,
-  NInput,
-  NButton,
-  NScrollbar,
-  useMessage,
-} from "naive-ui";
-import { reactive, ref, onMounted } from "vue";
+import { NScrollbar } from "naive-ui";
+import { onMounted } from "vue";
 import draggable from "vuedraggable";
 
-const { addTask, getTaskByColumn, loadTasks, persistColumns } = useTasksStore();
+import { useTasksStore } from "../stores/TasksStore";
+import AddTaskForm from "./AddTaskForm.vue";
 
-const message = useMessage();
-const formValue = reactive({
-  task: "",
-});
-
-const formRef = ref(null);
+const { getTaskByColumn, loadTasks, persistColumns } = useTasksStore();
 
 const columns = [
   {
@@ -70,32 +41,7 @@ const columns = [
   },
 ];
 
-const rules = {
-  task: {
-    required: true,
-    message: "Please enter your task",
-    trigger: ["input"],
-  },
-};
-
-function handleSubmit(e) {
-  e.preventDefault();
-  formRef.value
-    ?.validate()
-    .then(() => {
-      addTask("todo", {
-        title: formValue.task,
-      });
-      formValue.task = "";
-      message.success("New task added");
-    })
-    .catch((errors) => {
-      console.log(errors);
-      message.error("A task must not be empty");
-    });
-}
-
-function log() {
+function onChange() {
   persistColumns();
 }
 
@@ -105,11 +51,6 @@ onMounted(() => {
 </script>
 
 <style scoped>
-.form-wrapper {
-  padding: 1rem;
-  margin: auto;
-}
-
 .columns-wrapper {
   display: flex;
   gap: 3em;
